Ensure that the LEDs are never all off.
We introduce signals that each Colour state machine sends to Flashy to
tell when it's turned on or off.  Flashy keeps track of that and
immediately starts another flash when everything is off.  It starts
another flash by sending itself a signal, which allows us to keep
track of on and off LEDs in one state handler, but have the correct
state handler told to turn on a LED.

Signed-off-by: Russell Steicke <russells@adelie.cx>